backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.8.0
python-docx>=1.1.0
docxtpl>=0.16.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
openai>=1.0.0
dashscope>=1.14.0
pypinyin>=0.50.0
ruff>=0.1.0
mypy>=1.8.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
pyinstaller>=6.0.0
memorisdk>=0.1.0
